middleware/file: tree does ent checks
When inserting and deleting check for empty-non-terminals. This
adds the checks and functions for this to work.
