+++ middleware/file/zone_test.go
@@ -28,29 +28,3 @@ func TestNameFromRight(t *testing.T) {
 		}
 	}
 }
-
-func TestIsNonTerminal(t *testing.T) {
-	z := NewZone("example.org.", "stdin")
-
-	tests := []struct {
-		in       string
-		terminal bool
-		expected []string
-	}{
-		{"b.example.org.", false, nil},
-		{"a.b.example.org.", true, []string{"b.example.org."}},
-		{"a.b.c.example.org.", true, []string{"b.c.example.org.", "c.example.org."}},
-	}
-
-	for i, tc := range tests {
-		got, terminal := z.isNonTerminal(tc.in)
-		for j, g := range got {
-			if g != tc.expected[j] {
-				t.Errorf("Test %d: expected %s, got %s\n", i, tc.expected[j], g)
-			}
-		}
-		if terminal != tc.terminal {
-			t.Errorf("Test %d: expected terminal to be %t, got %t\n", i, tc.terminal, terminal)
-		}
-	}
-}
